+++ packages/react-craft-ai-decision-tree/src/components/decisionTreeWithPanel.jsx
@@ -1,24 +1,29 @@
 import ContainerDimensions from 'react-container-dimensions';
+import NodeInformations from './nodeInformation/nodeInformation';
 import PropTypes from 'prop-types';
-import React from 'react';
 import semver from 'semver';
 import styled from '@emotion/styled';
 import Tree from './tree';
+import React, { useState } from 'react';
 
 const DecisionTreeContainer = styled('div')`
   height: 100%;
 `;
 
-const DecisionTree = ({
+const DecisionTreeWithPanel = ({
   scale,
   position,
   data,
   height,
   width,
-  updateSelectedNode,
   updatePositionAndZoom
 }) => {
+  const [selectedNode, setSelectedNode] = useState('');
   const treeVersion = semver.major(data._version);
+
+  const closeNodeInformation = () => {
+    setSelectedNode(undefined);
+  };
 
   return (
     <DecisionTreeContainer
@@ -28,11 +33,16 @@ const DecisionTree = ({
         width: width
       }}
     >
+      <NodeInformations
+        tree={ data }
+        selectedNodePath={ selectedNode }
+        closeNodeInformation={ closeNodeInformation }
+      />
       <ContainerDimensions>
         {({ height, width }) => (
           <Tree
             version={ treeVersion }
-            updateSelectedNode={ updateSelectedNode }
+            updateSelectedNode={ setSelectedNode }
             height={ height }
             width={ width }
             position={ position }
@@ -47,22 +57,20 @@ const DecisionTree = ({
   );
 };
 
-DecisionTree.defaultProps = {
+DecisionTreeWithPanel.defaultProps = {
   position: [0, 0],
   scale: -1,
   updatePositionAndZoom: null,
-  auto: false,
-  updateSelectedNode: undefined
+  auto: false
 };
 
-DecisionTree.propTypes = {
+DecisionTreeWithPanel.propTypes = {
   scale: PropTypes.number,
   position: PropTypes.array,
-  updateSelectedNode: PropTypes.func,
   data: PropTypes.object.isRequired,
   height: PropTypes.number,
   width: PropTypes.number,
   updatePositionAndZoom: PropTypes.func
 };
 
-export default DecisionTree;
+export default DecisionTreeWithPanel;