+++ packages/react-craft-ai-decision-tree/src/components/edges.jsx
@@ -1,14 +1,20 @@
-import { COLOR_EDGES } from '../utils/constants';
 import { css } from 'react-emotion';
 import { select as d3Select } from 'd3';
 import PropTypes from 'prop-types';
 import React from 'react';
+import { DEFAULT_COLOR_EDGES, SELECTED_COLOR_EDGES } from '../utils/constants';
 
 // make links css rules
-const linksCssClass = css`
+const defaultLinksCssClass = css`
   fill: none;
-  stroke: ${COLOR_EDGES};
+  stroke: ${DEFAULT_COLOR_EDGES};
   stroke-width: 1.5;
+`;
+
+const selectedLinksCssClass = css`
+  fill: none;
+  stroke: ${SELECTED_COLOR_EDGES};
+  stroke-width: 2;
 `;
 
 class Edges extends React.Component {
@@ -18,7 +24,8 @@ class Edges extends React.Component {
       this.props.treeData,
       this.svgTreeRef,
       this.props.nodes,
-      this.props.links
+      this.props.links,
+      this.props.edgePath
     );
   }
 
@@ -28,7 +35,8 @@ class Edges extends React.Component {
       nextProps.treeData,
       this.svgTreeRef,
       nextProps.nodes,
-      nextProps.links
+      nextProps.links,
+      nextProps.edgePath
     );
 
     // Do not allow react to render the component on prop change
@@ -47,7 +55,7 @@ class Edges extends React.Component {
     this.svgTreeRef = input;
   };
 
-  renderTree = (treeData, svgDomNode, nodes, links) => {
+  renderTree = (treeData, svgDomNode, nodes, links, edgePath) => {
     // Cleans up the SVG on re-render
     d3Select(svgDomNode)
       .selectAll('*')
@@ -58,11 +66,9 @@ class Edges extends React.Component {
       .attr('height', this.props.height)
       .append('g');
 
-    let i = 0;
-
     // Update the nodes…
     let node = svg.selectAll('g.node')
-      .data(nodes, (d) => d.id || (d.id = ++i));
+      .data(nodes, (d) => d.id);
 
     // Enter any new nodes at the parent's previous position.
     let nodeEnter = node
@@ -80,15 +86,27 @@ class Edges extends React.Component {
       .attr('transform', (d) => `translate(${d.x},${d.y})`)
       .remove();
 
-    // Update the links…
+    // Update the links… (they are ordered)
     let link = svg.selectAll('path.link')
-      .data(links, (d) => d.target.id);
+      .data(links, (d) => {
+        d.linkClass = defaultLinksCssClass;
+        if (edgePath.indexOf(d.target.id) !== -1) {
+          d.linkClass = selectedLinksCssClass;
+        }
+        return d.target.id;
+      });
 
     // Enter any new links at the parent's previous position.
     link
       .enter()
       .insert('path', 'g')
-      .attr('class', linksCssClass)
+      .attr(
+        'class',
+        (d) =>
+          `${d.linkClass} ${
+            d.linkClass == selectedLinksCssClass ? 'selected-link' : ''
+          }`
+      )
       .attr('d', this.diagonal);
 
     // Transition exiting nodes to the parent's new position.
@@ -107,6 +125,7 @@ class Edges extends React.Component {
 }
 
 Edges.propTypes = {
+  edgePath: PropTypes.array.isRequired,
   treeData: PropTypes.object.isRequired,
   nodes: PropTypes.array.isRequired,
   links: PropTypes.array.isRequired,