+++ packages/react-craft-ai-decision-tree/src/index.js
@@ -1,4 +1,20 @@
+import DecisionRules from './components/nodeInformation/decisionRules';
 import DecisionTree from './components/decisionTree';
+import DecisionTreeWithPanel from './components/decisionTreeWithPanel';
+import NodeInformation from './components/nodeInformation/nodeInformation';
+import Prediction from './components/nodeInformation/prediction';
 import ReactDOM from 'react-dom';
+import Split from './components/nodeInformation/split';
+import Statistics from './components/nodeInformation/statistics';
 
-export default DecisionTree;
+export {
+  DecisionTreeWithPanel,
+  DecisionRules,
+  DecisionTree,
+  NodeInformation,
+  Prediction,
+  Statistics,
+  Split
+};
+
+export default DecisionTreeWithPanel;