+++ packages/react-craft-ai-decision-tree/src/utils/constants.js
@@ -1,5 +1,5 @@
 // Zoom
-export const ZOOM_EXTENT = [0.01, 1.5];
+export const ZOOM_EXTENT = [0.3, 1.5];
 
 // Fit to screen
 export const MARGIN = {
@@ -17,9 +17,14 @@ export const NODE_DEPTH = 100;
 
 // Color of the nodes and edges
 export const COLOR_EDGES_CAPTION_BG = 'rgba(255, 255, 255, 0.5)';
-export const COLOR_EDGES = '#ddd';
+export const SELECTED_COLOR_EDGES = '#FF7F00';
+export const DEFAULT_COLOR_EDGES = '#ddd';
 export const COLOR_NODES = '#ddd';
+export const SELECTED_BORDER_WIDTH = 5;
 
 // Color of the leaves
 export const COLOR_LEAVES_CONFIDENCE_0 = '#FEF3C2';
 export const COLOR_LEAVES_CONFIDENCE_1 = '#FFBC00';
+
+// Node path
+export const NODE_PATH_SEPARATOR = '-';