+++ CHANGELOG.md
@@ -5,15 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.18...HEAD) ##
+## [Unreleased](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.18...HEAD)
 
-## [0.0.18](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.17...v0.0.18) - 2019-02-28 ##
+## Added
+
+- Display the number of samples on the leaves
+- Add the possibility to import separate component:
+  - DecisionTree with panel by default
+  - DecisionTree alone
+  - Component of the panel (NodeInformation, Split, DecisionRules, Prediction and Statistics)
+
+## Updated
+
+- Change the popover of the leaf to a panel
+- Refactor the visualization of the data in all the nodes
+
+## [0.0.18](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.17...v0.0.18) - 2019-02-28
 
 ## Added
 
 - Expose the scale and the position of the tree.
 
-## [0.0.17](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.16...v0.0.17) - 2019-01-03 ##
+## [0.0.17](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.16...v0.0.17) - 2019-01-03
 
 ## Fixed